requirements.txt
feedparser==6.0.11
google-genai>=1.0.0
python-telegram-bot==21.9
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pytz==2024.2
